fix: address Kubernetes executor review findings
