+++ apps/web/src/locales/en/executors.json
@@ -209,7 +209,7 @@
   "kubernetesAccessModesHelp": "Select the access modes requested for each managed claim.",
   "kubernetesActiveSessions": "Active sessions",
   "kubernetesActiveSessionsDescription": "Sanitized Pod and workspace status for sessions using this executor. Refreshes every 90 seconds.",
-  "kubernetesActiveSessionsGuidance": "Stop preserves Kubernetes resources. Archive or Delete can remove a managed workspace. Requests show configured main-container CPU and memory, not actual use.",
+  "kubernetesActiveSessionsGuidance": "Stop preserves Kubernetes resources. Archive or Delete can remove a managed workspace. Existing claims are not deleted by Kandev. Requests show configured main-container CPU and memory, not actual use.",
   "kubernetesAdminSaveOnly": "Only an administrator can save Kubernetes settings.",
   "kubernetesAdminTestOnly": "Only an administrator can run this Kubernetes test.",
   "kubernetesAuthInCluster": "In-cluster service account",
@@ -260,7 +260,7 @@
   "kubernetesNamespaceRequired": "A namespace is required.",
   "kubernetesNoActiveSessions": "No active Kubernetes sessions.",
   "kubernetesMainContainerRequests": "Main-container requests",
-  "kubernetesOpenTask": "Open task {{task}}",
+  "kubernetesOpenTask": "Open task {{task}}, session {{session}}",
   "kubernetesPlatform": "Platform",
   "kubernetesPlatformHelp": "Adds the matching Kubernetes architecture selector to every Pod.",
   "kubernetesPod": "Pod",
