+++ apps/web/src/locales/pseudo/executors.json
@@ -209,7 +209,7 @@
   "kubernetesAccessModesHelp": "Śēĺēćţ ţĥē àććēśś ḿōďēś ŕēqũēśţēď ƒōŕ ēàćĥ ḿàńàĝēď ćĺàĩḿ.",
   "kubernetesActiveSessions": "Àćţĩvē śēśśĩōńś",
   "kubernetesActiveSessionsDescription": "Śàńĩţĩźēď Ƥōď àńď ŵōŕķśƥàćē śţàţũś ƒōŕ śēśśĩōńś ũśĩńĝ ţĥĩś ēxēćũţōŕ. Ŕēƒŕēśĥēś ēvēŕŷ 90 śēćōńďś.",
-  "kubernetesActiveSessionsGuidance": "Śţōƥ ƥŕēśēŕvēś Ķũƀēŕńēţēś ŕēśōũŕćēś. Àŕćĥĩvē ōŕ Ďēĺēţē ćàń ŕēḿōvē à ḿàńàĝēď ŵōŕķśƥàćē. Ŕēqũēśţś śĥōŵ ćōńƒĩĝũŕēď ḿàĩń-ćōńţàĩńēŕ ĆƤŨ àńď ḿēḿōŕŷ, ńōţ àćţũàĺ ũśē.",
+  "kubernetesActiveSessionsGuidance": "Śţōƥ ƥŕēśēŕvēś Ķũƀēŕńēţēś ŕēśōũŕćēś. Àŕćĥĩvē ōŕ Ďēĺēţē ćàń ŕēḿōvē à ḿàńàĝēď ŵōŕķśƥàćē. Ēxĩśţĩńĝ ćĺàĩḿś àŕē ńōţ ďēĺēţēď ƀŷ Ķàńďēv. Ŕēqũēśţś śĥōŵ ćōńƒĩĝũŕēď ḿàĩń-ćōńţàĩńēŕ ĆƤŨ àńď ḿēḿōŕŷ, ńōţ àćţũàĺ ũśē.",
   "kubernetesAdminSaveOnly": "Ōńĺŷ àń àďḿĩńĩśţŕàţōŕ ćàń śàvē Ķũƀēŕńēţēś śēţţĩńĝś.",
   "kubernetesAdminTestOnly": "Ōńĺŷ àń àďḿĩńĩśţŕàţōŕ ćàń ŕũń ţĥĩś Ķũƀēŕńēţēś ţēśţ.",
   "kubernetesAuthInCluster": "Ĩń-ćĺũśţēŕ śēŕvĩćē àććōũńţ",
@@ -260,7 +260,7 @@
   "kubernetesNamespaceRequired": "À ńàḿēśƥàćē ĩś ŕēqũĩŕēď.",
   "kubernetesNoActiveSessions": "Ńō àćţĩvē Ķũƀēŕńēţēś śēśśĩōńś.",
   "kubernetesMainContainerRequests": "Ḿàĩń-ćōńţàĩńēŕ ŕēqũēśţś",
-  "kubernetesOpenTask": "Ōƥēń ţàśķ {{task}}",
+  "kubernetesOpenTask": "Ōƥēń ţàśķ {{task}}, śēśśĩōń {{session}}",
   "kubernetesPlatform": "Ƥĺàţƒōŕḿ",
   "kubernetesPlatformHelp": "Àďďś ţĥē ḿàţćĥĩńĝ Ķũƀēŕńēţēś àŕćĥĩţēćţũŕē śēĺēćţōŕ ţō ēvēŕŷ Ƥōď.",
   "kubernetesPod": "Ƥōď",
