+++ apps/web/src/locales/pt-pt/executors.json
@@ -193,7 +193,7 @@
   "kubernetesAccessModesHelp": "Selecione os modos de acesso pedidos para cada claim gerida.",
   "kubernetesActiveSessions": "Sessões ativas",
   "kubernetesActiveSessionsDescription": "Estado sanitizado do Pod e do espaço de trabalho para as sessões que usam este executor. Atualiza a cada 90 segundos.",
-  "kubernetesActiveSessionsGuidance": "Parar mantém os recursos do Kubernetes. Arquivar ou eliminar pode remover um espaço de trabalho gerido. Os pedidos mostram CPU e memória configurados para o contentor principal, não o uso real.",
+  "kubernetesActiveSessionsGuidance": "Parar mantém os recursos do Kubernetes. Arquivar ou eliminar pode remover um espaço de trabalho gerido. Os claims existentes não são eliminados pelo Kandev. Os pedidos mostram CPU e memória configurados para o contentor principal, não o uso real.",
   "kubernetesAdminSaveOnly": "Apenas um administrador pode guardar definições do Kubernetes.",
   "kubernetesAdminTestOnly": "Apenas um administrador pode executar este teste do Kubernetes.",
   "kubernetesAuthInCluster": "Conta de serviço no cluster",
@@ -244,7 +244,7 @@
   "kubernetesNamespaceRequired": "É necessário um namespace.",
   "kubernetesNoActiveSessions": "Não há sessões Kubernetes ativas.",
   "kubernetesMainContainerRequests": "Pedidos do contentor principal",
-  "kubernetesOpenTask": "Abrir tarefa {{task}}",
+  "kubernetesOpenTask": "Abrir tarefa {{task}}, sessão {{session}}",
   "kubernetesPlatform": "Plataforma",
   "kubernetesPlatformHelp": "Adiciona a cada Pod o seletor de arquitetura Kubernetes correspondente.",
   "kubernetesPod": "Pod",
