+++ apps/web/src/locales/zh-cn/executors.json
@@ -193,7 +193,7 @@
   "kubernetesAccessModesHelp": "选择每个托管存储声明请求的访问模式。",
   "kubernetesActiveSessions": "活动会话",
   "kubernetesActiveSessionsDescription": "使用此执行器的会话所对应的 Pod 和工作区安全状态。每 90 秒刷新一次。",
-  "kubernetesActiveSessionsGuidance": "停止会保留 Kubernetes 资源。归档或删除可能会移除受 Kandev 管理的工作区。请求值显示主容器配置的 CPU 和内存，不是实际使用量。",
+  "kubernetesActiveSessionsGuidance": "停止会保留 Kubernetes 资源。归档或删除可能会移除受 Kandev 管理的工作区。Kandev 不会删除现有的声明。请求值显示主容器配置的 CPU 和内存，不是实际使用量。",
   "kubernetesAdminSaveOnly": "只有管理员可以保存 Kubernetes 设置。",
   "kubernetesAdminTestOnly": "只有管理员可以运行此 Kubernetes 测试。",
   "kubernetesAuthInCluster": "集群内服务账号",
@@ -244,7 +244,7 @@
   "kubernetesNamespaceRequired": "必须填写命名空间。",
   "kubernetesNoActiveSessions": "没有活动的 Kubernetes 会话。",
   "kubernetesMainContainerRequests": "主容器资源请求",
-  "kubernetesOpenTask": "打开任务 {{task}}",
+  "kubernetesOpenTask": "打开任务 {{task}}，会话 {{session}}",
   "kubernetesPlatform": "平台",
   "kubernetesPlatformHelp": "为每个 Pod 添加匹配的 Kubernetes 架构选择器。",
   "kubernetesPod": "Pod",
