+++ apps/web/src/locales/zh-hk/executors.json
@@ -193,7 +193,7 @@
   "kubernetesAccessModesHelp": "選擇每個託管存儲聲明請求的存取模式。",
   "kubernetesActiveSessions": "活動工作階段",
   "kubernetesActiveSessionsDescription": "使用此執行器的工作階段所對應的 Pod 和工作區安全狀態。每 90 秒刷新一次。",
-  "kubernetesActiveSessionsGuidance": "停止會保留 Kubernetes 資源。歸檔或刪除可能會移除受 Kandev 管理的工作區。請求值顯示主容器設定的 CPU 和記憶體，不是實際使用量。",
+  "kubernetesActiveSessionsGuidance": "停止會保留 Kubernetes 資源。歸檔或刪除可能會移除受 Kandev 管理的工作區。Kandev 不會刪除現有的聲明。請求值顯示主容器設定的 CPU 和記憶體，不是實際使用量。",
   "kubernetesAdminSaveOnly": "只有管理員可以儲存 Kubernetes 設定。",
   "kubernetesAdminTestOnly": "只有管理員可以執行此 Kubernetes 測試。",
   "kubernetesAuthInCluster": "集羣內服務帳戶",
@@ -244,7 +244,7 @@
   "kubernetesNamespaceRequired": "必須填寫命名空間。",
   "kubernetesNoActiveSessions": "沒有活動的 Kubernetes 工作階段。",
   "kubernetesMainContainerRequests": "主容器資源請求",
-  "kubernetesOpenTask": "開啟任務 {{task}}",
+  "kubernetesOpenTask": "開啟任務 {{task}}，工作階段 {{session}}",
   "kubernetesPlatform": "平台",
   "kubernetesPlatformHelp": "為每個 Pod 新增匹配的 Kubernetes 架構選擇器。",
   "kubernetesPod": "Pod",
